requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cssselect2==0.5.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.3
django-admin-honeypot==1.1.0
django-braces==1.15.0
django-ckeditor==6.2.0
django-countries==7.3.2
django-debug-toolbar==3.2.4
django-etc==1.3.1
django-filter==21.1
django-forms-dynamic==1.0.0
django-hitcount==1.3.5
django-htmx==1.9.0
django-import-export==2.7.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
django-star-ratings==0.9.2
django-storages==1.12.3
django-taggit==2.1.0
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
hashring==1.5.1
html5lib==1.1
idna==3.3
jsonfield==3.1.0
lxml==4.8.0
MarkupPy==1.14
memcache==0.11.0
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.45
Pillow==9.0.1
PyPDF3==1.0.6
PyQRCode==1.2.1
python-bidi==0.4.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
ratelim==0.1.6
reportlab==3.6.8
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
swapper==1.3.0
tablib==3.2.0
tinycss2==1.1.1
tqdm==4.63.0
typing_extensions==4.1.1
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.6
xlrd==2.0.1
xlwt==1.3.0
